Modularize editing dialog into component
