+++ app/containers/Admin/ReportingYear/ReportingYearTable.tsx
@@ -1,14 +1,10 @@
 import React, {useState} from 'react';
-import {Table, Modal, Container, Button} from 'react-bootstrap';
-import {JSONSchema6} from 'json-schema';
-import JsonSchemaForm from 'react-jsonschema-form';
-import FormObjectFieldTemplate from 'containers/Forms/FormObjectFieldTemplate';
-import FormFieldTemplate from 'containers/Forms/FormFieldTemplate';
+import {Table, Button} from 'react-bootstrap';
 import moment from 'moment-timezone';
 import {graphql, createFragmentContainer, RelayProp} from 'react-relay';
 import {ReportingYearTable_query} from '__generated__/ReportingYearTable_query.graphql';
-import reportingYearSchema from './reporting_year.json';
 import updateReportingYearMutation from 'mutations/reporting_year/updateReportingYearMutation';
+import ReportingYearFormDialog from './ReportingYearFormDialog';
 
 interface Props {
   relay: RelayProp;
@@ -32,45 +28,29 @@ function localeDateTimeToISO(date, time) {
 }
 
 export const ReportingYearTableComponent: React.FunctionComponent<Props> = props => {
-  const initialFormState = {
-    editingYear: null,
-    formFields: null
-  };
-  const [state, setState] = useState(initialFormState);
+  const [editingYear, setEditingYear] = useState(null);
 
   const {query, relay} = props;
   if (!query.allReportingYears || !query.allReportingYears.edges) {
     return <div />;
   }
 
-  const clearForm = () => {
-    setState(prev => {
-      return {
-        ...prev,
-        ...initialFormState
-      }
-    });
-  }
+  const clearForm = () => setEditingYear(null);
 
-  // TODO: Refactor / implement this in a better way
-  const editYear = node => {
-    setState((prev) => {
-      return {
-        ...prev,
-        editingYear: node,
-        formFields: {
-          applicationOpenDate: isoToLocaleDate(node.applicationOpenTime),
-          applicationOpenTime: isoToLocaleTime(node.applicationOpenTime),
-          applicationCloseDate: isoToLocaleDate(node.applicationCloseTime),
-          applicationCloseTime: isoToLocaleTime(node.applicationCloseTime),
-          applicationResponseTime: isoToLocaleDate(node.applicationResponseTime)
-        }
-      }
-    });
+  const editYear = node => setEditingYear(node);
+
+  const displayFormFields = (node) => {
+    if(!node) return null
+    return {
+      applicationOpenDate: isoToLocaleDate(node.applicationOpenTime),
+      applicationOpenTime: isoToLocaleTime(node.applicationOpenTime),
+      applicationCloseDate: isoToLocaleDate(node.applicationCloseTime),
+      applicationCloseTime: isoToLocaleTime(node.applicationCloseTime),
+      applicationResponseTime: isoToLocaleDate(node.applicationResponseTime)
+    }
   };
 
-  // TODO: Refactor / implement this in a better way
-  const saveReportingYear = async ({formData}) => {
+  const saveReportingYear = ({formData}) => {
     const finalData = {
       applicationOpenTime: localeDateTimeToISO(
         formData.applicationOpenDate,
@@ -88,7 +68,7 @@ export const ReportingYearTableComponent: React.FunctionComponent<Props> = props
 
     updateReportingYearMutation(relay.environment, {
       input: {
-        id: state.editingYear.id,
+        id: editingYear.id,
         reportingYearPatch: {
           ...finalData
         }
@@ -96,40 +76,6 @@ export const ReportingYearTableComponent: React.FunctionComponent<Props> = props
     })
     .then(clearForm);
   };
-
-  const editModal = (
-    <Modal
-      centered
-      size="xl"
-      show={Boolean(state.editingYear)}
-      onHide={clearForm}
-    >
-      <Modal.Header closeButton>
-        <Modal.Title>
-          Edit Reporting Year {state.editingYear?.reportingYear}
-        </Modal.Title>
-      </Modal.Header>
-      <Modal.Body>
-        <Container>
-          <JsonSchemaForm
-            omitExtraData
-            liveOmit
-            schema={reportingYearSchema.schema as JSONSchema6}
-            uiSchema={reportingYearSchema.uiSchema}
-            formData={state.formFields}
-            FieldTemplate={FormFieldTemplate}
-            ObjectFieldTemplate={FormObjectFieldTemplate}
-            showErrorList={false}
-            onSubmit={saveReportingYear}
-          >
-            <Button type="submit" variant="primary">
-              Save Reporting Year
-            </Button>
-          </JsonSchemaForm>
-        </Container>
-      </Modal.Body>
-    </Modal>
-  );
 
   return (
     <>
@@ -163,7 +109,13 @@ export const ReportingYearTableComponent: React.FunctionComponent<Props> = props
           })}
         </tbody>
       </Table>
-      {editModal}
+      <ReportingYearFormDialog
+        show={Boolean(editingYear)}
+        year={editingYear?.reportingYear}
+        formFields={displayFormFields(editingYear)}
+        clearForm={clearForm}
+        saveReportingYear={saveReportingYear}
+      />
     </>
   );
 };
