+++ app/pages/admin/reporting-years.tsx
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import {graphql} from 'react-relay';
 import {reportingYearsQueryResponse} from 'reportingYearsQuery.graphql';
 import DefaultLayout from 'layouts/default-layout';
-import ReportingYearTable from 'containers/Admin/ReportingYearTable';
+import ReportingYearTable from 'containers/Admin/ReportingYear/ReportingYearTable';
 import {ADMIN_GROUP} from 'data/group-constants';
 
 const ALLOWED_GROUPS = ADMIN_GROUP;
